Moves Metrics Filter to separate auto-config
fixes gh-457
