+++ spring-cloud-gateway-core/src/main/resources/META-INF/spring.factories
@@ -3,5 +3,6 @@ org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
 org.springframework.cloud.gateway.config.GatewayClassPathWarningAutoConfiguration,\
 org.springframework.cloud.gateway.config.GatewayAutoConfiguration,\
 org.springframework.cloud.gateway.config.GatewayLoadBalancerClientAutoConfiguration,\
+org.springframework.cloud.gateway.config.GatewayMetricsAutoConfiguration,\
 org.springframework.cloud.gateway.config.GatewayRedisAutoConfiguration,\
 org.springframework.cloud.gateway.discovery.GatewayDiscoveryClientAutoConfiguration
